Adds a non-insitu bandaid for render queries with subqueries in distributed
* Adds a temporary NonInsituQueryClassification specialization for
distributed render queries which scans all nodes of the DAG for aggregates
or subqueries

* Passes InSituFlags through the RenderQueryStructure object of the
RenderQueryParseData and used as input for poly column injection.

* Removes the RenderQueryRunnerAgg::isAggregateQuery logic and instead
relies on the InSituFlags of the current render_info_ to deterine
whether to run the query on the aggregator (non-insitu) or push to the
leaves (insitu).

* Hoists the RelAlgDag build out of RenderQueryRunner::validateQuery()
& executeQuery() methods and into a statically-defined
RenderQueryRunner::buildRelAlgDagAndClassifyRender() method instead. Now
the dag build and non-insitu classification is done immediately after
calcite parse.