Skip refresh for append foreign tables with no datawrapper
Followup to the original item where we add support for append workloads (which took a different path).

The changes here involve:

skipping the refresh if there is no datawrapper available for the foreign table in question
Changing the test MockDataWrapper system so that a mock wrapper that is "empty" (does not wrap around an existing wrapper) does not count as an existing wrapper (so that we don't consider an empty mock wrapper to be a valid wrapper). This was needed to prevent the tests from automatically refreshing because we had added a mock wrapper before any proper wrapper had existed.