parquet import respect number of threads param
Updated the parquet import path to respect the number of threads copy param.

The other data wrappers that use this parameter do so via FileBufferParser, and perform their own copy parameter validation. Since the parquet wrapper does not do this, it has been updated to pass in this parameter directly when creating the temporary wrapper (small change since it already used portions of the param to determine data wrapper type).

We currently have no tests set up to validate the number of threads that we create as part of an import, so this change was tested manually to confirm that the parameter is now being recognized and used when calculating the number of threads to use within the DW.