Turn calcite request timeout crash into an exception.
Currently, it is possible for a crash to occur when a schema write operation (e.g. column rename) runs for a long time while a select query is issued concurrently for the same table. This occurs because, as part of a select query, a request is made to Calcite, which makes a request to the get_internal_table_details_for_database Thrift API. The get_internal_table_details_for_database API acquires a read schema lock for the table. If schema lock acquisition takes more than 5 seconds/configured Calcite service timeout, the request to Calcite throws a timeout exception that results in a crash.

This change is to prevent this crash by detecting a Calcite request timeout and changing it into an exception, so that the query will still fail, but will no longer crash just because the request is timing out. A further change may detect the specific reason for a timeout and can deliver a custom exception message mentioning that the table is locked.