Added FileInfo unit tests
Added unit tests for FileInfo API. This work is part of an ongoing item to refactor and update the FileMgr subsystem and these unit tests should help us be more confident there are no breaks along the way.

In order to support test-comparison data files that are of a reasonable size, the FileMgr system was changed to use an adjustable metadataPageSize value (set at FileMgr construction) rather than a fixed METADATA_PAGE_SIZE constant. This enables us to use files of a smaller size than default metadata file sizes. The code-paths for setting page/metadata_page sizes were also cleaned up so that this info is only passed via the GlobalFileMgr/FileMgr so as to disallow mismatches between the various levels (PageSize/MetadataPageSize is now set when constructing a GFM, which is reused when setting sizes in child FileMgrs rather than each setting it individually).

The METADATA_PAGE_SIZE constant has been replaced with DEFAULT_METADATA_PAGE_SIZE and the location has been consolidated with DEFAULT_PAGE_SIZE.

Const-correctness was updated in a few locations.