Change error message for max precision
Fix error messaging when creating a table to report the documented maximum precision (18) and adjusted tests to match. Added named constants for our maximum numeric precision in sql_types.h.

When performing arithmetic or casts it is possible to use 19-digit precision (for some values that we can represent) but this limit is not documented and should not be the one shown to a user when they attempt to create a type that is too precise.