core load balancer testing (#6784)
Basic testing for the load balancer system.

This PR introduces a test suite that will create and manage multiple HeavyDB servers that point to the same data directory using the --multi-instance option. Once these servers are created we perform a few basic tests to verify that changes to the system instigated by one server are properly reflected on all servers.

Currently there are failures that need to be addressed as followup PRs.