Added dictionary refresh to multi-instance table cache refresh (#6879)
Currently multi-instance cache refreshes do not refresh dictionaries when doing table-specific refreshes. This means that dictionary metadata can get out of sync between instances (notably with the number of dictionary entries if one server inserts a new string).

This change adds dictionary refreshes to the cache invalidation to make sure that we get up to date dictionary data in multi-instance mode.

Tests for this change have been added as part of a separate PR (The load-balancer functional tests). This change was pulled out of the test PR to isolate the functional changes for tracking purposes.

Signed-off-by: jack <jack@omnisci.com>