Update system dynamically when new license is set.
This PR fixes a few issues:

Fixes interaction with cmd line num sessions and license num sessions.
Currently when we update the maximum number of concurrent sessions we limit the value to the minimum of the cmd line value (passed via --num_sessions=... and the value in the license. However, "0" is considered unlimited, so a user can specify --num-sessions=0 and get unlimited sessions even if their license has a strict limit.
This is solved by reworking the comparison when setting number of sessions.

Fixes persistence of num sessions.
Currently there is no distinction between the number of sessions specified by the command line and the number of sessions persisted between restarts, so we had no way of telling if the restarted value came from a cmd line parameter or a previous license and which should take priority.
This is fixed by creating a new record of the number of sessions in the DBhandler so that we have a distinction between the current value, the cmd-line value, and the license value.

Dynamically updates the number of sessions when a new license is set.
We no longer need to restart the server. When a new license is set it will automatically disconnect all existing sessions and implement the new license limitations.
This is fixed by running some new update code when the DBHandler is created or when a new license is set via the thrift interface.

Fixes issue with license change not being respected at restart.
A license is now taken into account before creating the session store, so license changes will be persisted through a serer restart.
This is fixed by moving the sessions store initialization code to after the license parsing code.

Adds testing for both license parsing during reset, and dynamic setting using thrift interface.
Adds mocking support for the license to allow this testing.

Signed-off-by: jack <jack@omnisci.com>