Table Function to generate Contour Lines/Polygons (Part 1) (#7018)
Initial implementation of Table Functions to generate contour lines or polygons from raster inputs.

Simple version expects raster width/height to be provided, and the raster rows to be lon/lat axis-aligned and in sorted row-major order. This is unlikely to be the case with real-world data, hence...

Alternative version integrates `geo_rasterize` such that it can be provided with non-axis-aligned (e.g. satellite path) raster data, or even an arbitrary sparse 2D point-cloud, and it will re-rasterize at the given world space (meters) resolution. The other parameters of the rasterizing step are the same as for other TFs which use this (currently RF Prop).

Both versions have the following additional parameters:

* `raster` - a `CURSOR` input expecting `lon`/`lat` and `value` (each can be `float` or `double`)
* `contour_interval` - the spacing of the output lines or polygon regions in the same (same type as `value`)
* `contour_offset` - offset of contours (e.g. offset = 500, interval = 1000, produces lines at 500, 1500, 2500 etc.)
* `flip_latitude` - a hopefully-temporary debugging control that optionally inverts the resulting lines/polys

Lines can be rendered directly in a render query. Polys cannot yet.

Part 2 will provide:

* Renderable polys
* Tests
* Addressing remaining PR issues
* Hopefully the removal and hard-wiring of the `flip_latitude` parameter
* Any other required fixes based on FE development

Signed-off-by: jack <jack@omnisci.com>