Refactor UDTF implementations to prevent multiple symbol definitions errors (#7098)
This is part of an ongoing effort to reduce build time of UDTF related
files. Currently, heavydb generates a single cpp file containing all
the initialization and template instantiation for all UDTFs
declarations. While this works, it is pretty slow as the compiler has
to compile a single C++ file with thousands of lines with template
instantiation.

The aim is to split the content of the cpp file into multiple
files. However, during the implementation of this task, we discovered
that some UDTFs couldn't be compiled as symbols were being defined
multiple times in object files. This issue is being addressed in
this commit.

Signed-off-by: jack <jack@omnisci.com>