Updated datawrapper internal metadata cache when updating disk cache (#7115)
A segfault is occurring because some of the Import Buffers for the geo types are null. Specifically, the non-varlen types are assigned null buffers which we try to access.

This is occurring during the process in the cache refresh where we re-cache the pre-existing metadata (ForeignStorageCache::cacheMetadataVec()). In this function we check to see if the metadata we are caching is equal to the metadata already in the cache for a given chunk, and if it is different, then we erase the chunk from the cache. As we are not appending any new data, all the chunks are already in the cache, and yet when we request metadata from the data wrapper and attempt to store it in the cache it registers as different and clears the corresponding chunks for all varlen types:

cacheMetadataVec() for chunk: {1,4,14,1,1}
metadata parameter:
type: INTEGER[] numBytes: 0 numElements 1 min: 3 max: 3 has_nulls: 0
metadata in cache:
type: INTEGER[] numBytes: 8 numElements 1 min: 3 max: 3 has_nulls: 0
erasing chunk from cache...
It looks like the difference here is due to the metadata we request from the wrapper not having a specified numBytes, whereas the metadata stored in the cache has a known numBytes.

The reason why the numBytes is different is because the function AbstractTextFileDataWrapper::populateChunkMetadata() relies on running a metadata scan to get cached metadata. The function maintains a chunk_encoder_buffers_ object for storing and returning metadata which is preserved between calls to populateChunkMetadata(). When we attempt to get metadata we return values from chunk_encoder_buffers_ OR any data from cached buffers we obtained during the metadata scan. If no metadata scan was performed (as is the case if we skip the metadata scan due to the underlying files not changing) then we will return the numBytes based on the chunk_encoder_buffers_ because there are no cached buffers to use. Notably, when there are cached buffers, we don’t update the chunk_encoder_buffers_ so when we DO run the metadata scan, we can return one set of metadata, but save another set for reuse - hence the issue with mismatched metadata.

---------

Co-authored-by: Paul Aiyedun <paul.aiyedun@heavy.ai>
Signed-off-by: Misiu Godfrey <misiu.godfrey@kraken.mapd.com>