Moved read-only checks into FileMgr control and add unit tests (#7080)
Moved (almost all) read-only tests out of File.cpp and into the control of the FileMgr system. This was needed because we perform some file operations (namely through the CachingFileMgr) that are allowed in read-only mode which would currently fail, crashing the system.

Investigating the usage of the File.cpp functions, they are all (with the exception of renaming files) gated by the FileMgr system. The current design calls for the read-only checks to be hoisted into the FileMgr so that failing due to read-only mode is controlled by the FileMgr and/or CachingFileMgr.

Notable changes:

Read-only unit tests added for FileMgr, CachingFileMgr, FileBuffer, and FileInfo. This setup should also help for upcoming refactors of the FileMgr system.
Unifying the fileMgrVersion and db_version_ variables into a constant. We were currently maintaining multiple copies of what should have been a universal constant dictating the db version (should always be "2", until we need to update to a new version).
Added a wrapper function for CHECKs in the FileMgr subsystems to allow a throw on failure rather than a crash (controllable by global parameter). This allows much unit testing as we can now test for inputs that would have otherwise caused crashes and leads to better code-coverage.
Added gatekeeping functions in the FileMgr/CachingFileMgr to provide read-only sanitized access to the File.cpp functions.
Minor cleanup and refactoring.
Changed some FileMgr internals to use smart pointers instead of raw memory allocations to prevent memory leaks if constructors throw (which was added to the tests).

Signed-off-by: Misiu Godfrey <misiu.godfrey@kraken.mapd.com>