Server crashes when projecting geo column in some cases (#7166)
* Fix incorrect codegen for queries with column and column expression projections of the same column

There are parts of the codegen logic that conditionally generate code
based on whether a column is lazy fetched, but at the same time,
determination of lazy fetched columns is done as part of codegen.
So it is possible to end up in a state where previous codegen steps
assume that a column is lazy fetched while a later codegen step
determines that the column is not lazy fetched. In such a case,
the compilation step needs to be restarted in order to avoid
incorrectly generated code.

* Fix crash when RelAlgExecutionUnit is a null pointer

* Get rid of target_exprs from PlanState

* Refine the logic to keep mutual exclusiveness

---------

Co-authored-by: Paul Aiyedun <paul.aiyedun@heavy.ai>
Signed-off-by: Misiu Godfrey <misiu.godfrey@kraken.mapd.com>