Fix crash when accessing large ddl_statement column value
* Fix a crash that could occur when accessing the ddl_statement column
of the tables system table in cases where the column value contains a
DDL statement with length that exceeds the maximum string length for
string dictionary encoded columns.

* Add a new `TypedImportBuffer::addDictStringWithTruncation()` method
that truncates strings with length that exceed the maximum string
length for string dictionary encoded columns.

* Consistently use the new
`TypedImportBuffer::addDictStringWithTruncation()` method for string
values in the system table data wrappers.

Signed-off-by: Misiu Godfrey <misiu.godfrey@kraken.mapd.com>