Horizontal scaling concurrency testing (#7275)
Addresses a bug in Horizontal Scaling mode where the sqlite connection is not reset and therefore caches outdated information when a database is dropped and re-created.

Also moved most of the code in heavyai_locks.h into a new .cpp file in order to make re-compiles and debugging easier/faster.

Also fixed some other build breaks.

Signed-off-by: Misiu Godfrey <misiu.godfrey@kraken.mapd.com>