Added missing dependencies for tests and initheavy (#7318)
There are currently several dependencies that are needed to run initheavy/heavydb and most tests that are not connected in the makefiles (they are not needed to compile, but they are needed to run). This means that our typical workflow is to make all targets with tests disabled, and then enable tests and build them directly (or make all targets, which builds all tests).

In addition, most of our testing builds all tests and then only runs some subset. We can save some time/space if we create a new build target for these dependencies and then make sure we only build the tests that are actually running.

Signed-off-by: Misiu Godfrey <misiu.godfrey@kraken.mapd.com>