Fix Table Corruption Issue after encountering Uncheckpointed pages. (#7430)
When a new FileMgr is instantiated over existing data (i.e. when a new server instantiated on existing data), we will delete any pages on disk that where not checkpointed (i.e. delete any pages where the Page Epoch is greater than the last checkpointed FileMgr Epoch). This deletion process, however, works outside our normal process for page deletion and does not mark the deleted pages as "dirty" nor flushes them (it performs an fwrite() and otherwise relies on an fclose() call as part of the FileInfo's destructor to guarantee a sync to disk). This alternate process can lead to issues because the deletion of these pages is not guaranteed to flush to disk - even if the FileMgr writes other data to disk and increments it's epoch.

Specifically, a problem can arise when we "delete" uncheckpointed pages, then write new data, checkpoint - incrementing the FileMgr epoch, and then the server crashes without syncing the uncheckpointed pages. A checkpoint will only flush pages marked "dirty" to disk, so this does not include any uncheckpointed pages that were deleted. This means after checkpoint we can be in a state where the uncheckpointed pages have not been flushed to disk, but the new FileMgr epoch has, which means that the uncheckpointed pages are marked with an epoch that we now consider valid. If a crash occurs now, the uncheckpointed pages on disk will be read as checkpointed pages, even though they were never checkpointed (because the epoch incremented as part of an unrelated checkpoint).

This PR fixes this issue by having the deletion process for uncheckpointed pages (the freePageImmediate() function) mark the pages as dirty when it deletes them, thereby flushing them to disk during any subsequent checkpoint. It also moves towards standardizing the way we perform a lot of these page writes, making sure we acquire the appropriate locks (the FileMgr tends to skip locking during initialization as it assumes it won't be necessary).

Because of the complication of this test case (it requires setting up a particular state, crashing the system, then recovering in a particular way while relying on the OS's file IO buffer behaving a certain way), testing was performed manually.

Signed-off-by: Misiu Godfrey <misiu.godfrey@kraken.mapd.com>