Add default slab size server configuration parameter (#7651)
The max slab size configuration parameter currently determines both the
maximum possible and default slab allocation size. This update
introduces a configuration parameter that takes on the latter
responsibility of determining the default slab allocation size and
allows allocations above this size up to the configured max slab size.

* Add server configuration parameters for default CPU and GPU slab sizes.

* Set the default slab size to the same value as the max slab size when
it is not configured.

* Add new test cases for default slab size and update existing tests to
set this size.

* Update buffer pool and slab size related variable names to conform to
style guide.

* Rename current_max_slab_page_size_

* Add max/min CPU/GPU slab size validation

* Remove redundant parenthesis

* Use default_slab_size_ in DramArena constructor

* Move slab size calculation to a shared function

* Add slab count assertion to assertExpectedBufferMgrAttributes()

Signed-off-by: Misiu Godfrey <misiu.godfrey@kraken.mapd.com>