Optimize raster datawrapper (#7751)
This PR mainly focuses on optimizing the import process of the datawrapper. Key optimizations include:

Efficiently calculate point metadata by only processing points for a chunk's bounding box instead.
Use placeholder metadata for expensive (or useless) columns, like bands and angle column and point-type columns.
Remove ImportBuffers in favour of arrays of bytes.
Skip encoders for columns who's metadata are already known and write directly to the buffer. This will skip metadata updates and validation.
Lean towards performing operations on arrays of pixels instead of individual pixels for better compiler optimization of code.
Performing some loop optimizations that the compiler is missing.

More details:
Efficiently calculate point metadata by only processing points for a chunk's bounding box instead.

An expensive part of metadata calculation is processing all the point data to calculate the metadata for these columns. We can reduce the amount of values we need to process to obtain metadata by only calculating the min/max values for the bounding box of a chunk. Since these are lon/lat values, the min/max of each should be guaranteed to be in the bounding box, and therefore we can avoid performing the calculation for all pixels inside the chunk.

Use Placeholder Metadata:
If using kPOINT types for point columns, the metadata (beyond number of elements) is useless, so just use placeholder.
We also use placeholder for band data and angle data which can be refined later when data is actually processed.

Remove ImportBuffers:
Using importbuffers for these loads is quite slow. They have been removed and any necesary temporary buffers are represented with arrays of bytes. This was done with vectors at first, but the element constructor overhead was very high.

Skip Encoders:
For columns where we have already generated metadata (or the metadat does not matter) we don't have any need for metadata updates or validation, so it is much faster to write directly to the buffers.

Use Vector Operations:
Make use of array-based transformations when possible. I've also split up several loops that do multiple things into multiple more cohesive loops. While we now have to iterate over the data more often, these loops can often be vectorized because of their simplicity and we get better performance as a result.

Optimizations by hand:
A few cases were apparent where the compiler was not able to perform invariant hoisting or other loop optimizations. Some of this code was hand-optimized for performance speedups.

Signed-off-by: Misiu Godfrey <misiu.godfrey@kraken.mapd.com>