Added Raster support for FSI import including missing options (#7757)
Added support for raster import through the raster datawrapper.

This changes how we import raster tables. Not only do we now organize the data by tiles as we do in the raster data wrapper, but the same restrictions apply now. Notably, the schema currently needs to includ all the intended columns, and you can't filter out point columns. We have discussed changing this in a future release.

I've also added support for missing options, such as adding 'auto' for the block size, and adding band filtering and point transform specifications. This should support most of the options that classic raster import did, with the exception of the some error handling options like remove nulls rows and the maximum number of bad rows.

Signed-off-by: Misiu Godfrey <misiu.godfrey@kraken.mapd.com>