Add serialization support, file globbing support, and S3 support. (#7761)
Added file globbing by adding in the standard FSI file globbing plus mapping a separate RasterImporter per file.

Added serialization support. Not much needs to be done for this wrapper, so some of the basic maps were added.

Added S3 support. GDAL has native support for s3, so the main complexity is passing the authentication from the server and user_mappings to GDAL.

---------

Co-authored-by: Misiu Godfrey <misiu.godfrey@heavy.ai>
Signed-off-by: Misiu Godfrey <misiu.godfrey@kraken.mapd.com>