Make sure parquet append refresh compares schema to existing files (#7881)
Currently a parquet append refresh with multiple files will only validate the schema of the new file against other files that have been added during the append. This means there could be a schema mismatch between the new and old files and we would not detect the difference.

This change makes sure the schema is compared between the new and old files.

Signed-off-by: Misiu Godfrey <misiu.godfrey@kraken.mapd.com>