Raster Import Change Fragment Size (#7911)
The goal of this PR is to make sure we map raster tiles directly to fragments when we import. Essentially, the RasterDataWrapper is already fragmenting the data according to the user/file's tiling information, and that data grouping should be preserved. Currently an import will ignore the fragmentation performed by the Data Wrapper and pull all the data into the fragmenter, which re-fragments that data based on the existing table's requirements. This means that we will often combine the tile-oriented data into fewer large fragments that do not align geospatially and defets the purpose of the tileing.

Two main features added in this branch:

New PassThroughFragmenter which always assigns a new fragment when inserting a block of data. This means we can make sure that a raster import maintains a one-to-one mapping of raster tiles to fragments by creating a new fragment whenever a tile is imported, rather than appending new data to a previous non-full tile (as might be generated if we have a partially-full tile in a file).
Raster import will now alter the fragment size of the table it is imported into. This is necessary to force the one-to-one mapping of tiles to fragments. Checks are in place to make sure that the table is either empty, or already has the correct size, as we should not be able to alter the fragment size of a table that already contains data.

Signed-off-by: Misiu Godfrey <misiu.godfrey@kraken.mapd.com>