parquet point import (#8149)
Adds support to import two parquet double columns into a single heavy POINT column.
This is done by having the parquet data wrapper maintain a mapping of parquet columns to heavy columns, which can now support more than just one-to-one mappings (although it only supports the double/double -> point type at the moment).
This type-compression is checked at validation time and then the mapping is used in subsequent data-reading.

In addition:

Adds support for the LONLAT key to reverse the order of the lat/lon columns (same as CSV has).
Suports null entries in either double column by nullifying the POINT column.
chunk_metadata_map is now inherant to all data wrappers (most wrappers defined it already).
Created a new ForeignDataWrapperUnitTest file that contains a lot of the helper functions that were previously in RasterTableTests or ForeignTableDMLTests as this code is now used in both for unit testing.

Signed-off-by: Misiu Godfrey <misiu.godfrey@kraken.mapd.com>