Fixed s3 tests to target new public directories (#8136)
Added a new S3 bucket heavyai-import-test to replace the one that was deleted. Updated the tests to point at this new bucket.

The parquet files still throw an exception because the schema does not match up (I salvaged the existing files, so we are using the same ones as before). If we want these tests to do something other than check that filtering works we can add coverage, but there should not be a difference between them and the local versions.

Signed-off-by: Misiu Godfrey <misiu.godfrey@kraken.mapd.com>